Moved VS detection into new file
Removed VS 2015 code
